Got rid of MSVC
We are using msys2 with qtcreator and mingw now. Instructions how to compile under windows with msys2 will be added to the readme.md and we are planning to add appveyor as a windows ci, so you can look up in the appveyor.yml how our project is compiled under windows with msys2.